+++ ogn/parser/parse.py
@@ -2,7 +2,7 @@ import re
 from datetime import datetime
 
 from ogn.parser.utils import createTimestamp, parseAngle, kts2kmh, feet2m, fpm2ms
-from ogn.parser.pattern import PATTERN_APRS, PATTERN_RECEIVER_BEACON, PATTERN_AIRCRAFT_BEACON
+from ogn.parser.pattern import PATTERN_APRS_POSITION, PATTERN_APRS_STATUS, PATTERN_RECEIVER_BEACON, PATTERN_AIRCRAFT_BEACON
 from ogn.parser.exceptions import AprsParseError, OgnParseError
 
 
@@ -12,22 +12,30 @@ def parse_aprs(message, reference_date=None, reference_time=None):
         reference_date = now.date()
         reference_time = now.time()
 
-    match = re.search(PATTERN_APRS, message)
-    if match:
-        return {'name': match.group('callsign'),
-                'receiver_name': match.group('receiver'),
-                'dstcall': match.group('dstcall'),
-                'timestamp': createTimestamp(match.group('time'), reference_date, reference_time),
-                'latitude': parseAngle('0' + match.group('latitude') + (match.group('latitude_enhancement') or '0')) *
-                (-1 if match.group('latitude_sign') == 'S' else 1),
-                'symboltable': match.group('symbol_table'),
-                'longitude': parseAngle(match.group('longitude') + (match.group('longitude_enhancement') or '0')) *
-                (-1 if match.group('longitude_sign') == 'W' else 1),
-                'symbolcode': match.group('symbol'),
-                'track': int(match.group('course')) if match.group('course_extension') else 0,
-                'ground_speed': int(match.group('ground_speed')) * kts2kmh if match.group('ground_speed') else 0,
-                'altitude': int(match.group('altitude')) * feet2m,
-                'comment': match.group('comment')}
+    match_position = re.search(PATTERN_APRS_POSITION, message)
+    if match_position:
+        return {'name': match_position.group('callsign'),
+                'receiver_name': match_position.group('receiver'),
+                'dstcall': match_position.group('dstcall'),
+                'timestamp': createTimestamp(match_position.group('time'), reference_date, reference_time),
+                'latitude': parseAngle('0' + match_position.group('latitude') + (match_position.group('latitude_enhancement') or '0')) *
+                (-1 if match_position.group('latitude_sign') == 'S' else 1),
+                'symboltable': match_position.group('symbol_table'),
+                'longitude': parseAngle(match_position.group('longitude') + (match_position.group('longitude_enhancement') or '0')) *
+                (-1 if match_position.group('longitude_sign') == 'W' else 1),
+                'symbolcode': match_position.group('symbol'),
+                'track': int(match_position.group('course')) if match_position.group('course_extension') else 0,
+                'ground_speed': int(match_position.group('ground_speed')) * kts2kmh if match_position.group('ground_speed') else 0,
+                'altitude': int(match_position.group('altitude')) * feet2m,
+                'comment': match_position.group('comment')}
+
+    match_status = re.search(PATTERN_APRS_STATUS, message)
+    if match_status:
+        return {'name': match_status.group('callsign'),
+                'receiver_name': match_status.group('receiver'),
+                'dstcall': match_status.group('dstcall'),
+                'timestamp': createTimestamp(match_status.group('time'), reference_date, reference_time),
+                'comment': match_status.group('comment')}
 
     raise AprsParseError(message)
 
@@ -42,13 +50,14 @@ def parse_ogn_aircraft_beacon(aprs_comment):
                 'climb_rate': int(ac_match.group('climb_rate')) * fpm2ms,
                 'turn_rate': float(ac_match.group('turn_rate')),
                 'flightlevel': float(ac_match.group('flight_level')) if ac_match.group('flight_level') else None,
-                'signal_strength': float(ac_match.group('signal')),
+                'signal_quality': float(ac_match.group('signal_quality')),
                 'error_count': float(ac_match.group('errors')),
                 'frequency_offset': float(ac_match.group('frequency_offset')),
                 'gps_status': ac_match.group('gps_accuracy'),
                 'software_version': float(ac_match.group('flarm_software_version')) if ac_match.group('flarm_software_version') else None,
                 'hardware_version': int(ac_match.group('flarm_hardware_version'), 16) if ac_match.group('flarm_hardware_version') else None,
-                'real_address': ac_match.group('flarm_id')}
+                'real_address': ac_match.group('flarm_id'),
+                'signal_power': float(ac_match.group('signal_power')) if ac_match.group('signal_power') else None}
     else:
         return None
 
@@ -63,10 +72,19 @@ def parse_ogn_receiver_beacon(aprs_comment):
                 'total_ram': float(rec_match.group('ram_total')),
                 'ntp_error': float(rec_match.group('ntp_offset')),
                 'rt_crystal_correction': float(rec_match.group('ntp_correction')),
+                'voltage': float(rec_match.group('voltage')) if rec_match.group('voltage') else None,
+                'amperage': float(rec_match.group('amperage')) if rec_match.group('amperage') else None,
                 'cpu_temp': float(rec_match.group('cpu_temperature')) if rec_match.group('cpu_temperature') else None,
-                'rec_crystal_correction': int(rec_match.group('manual_correction')) if rec_match.group('manual_correction') else 0,
-                'rec_crystal_correction_fine': float(rec_match.group('automatic_correction')) if rec_match.group('automatic_correction') else 0.0,
-                'rec_input_noise': float(rec_match.group('input_noise')) if rec_match.group('input_noise') else None}
+                'senders_visible': int(rec_match.group('visible_senders')) if rec_match.group('visible_senders') else None,
+                'senders_total': int(rec_match.group('senders')) if rec_match.group('senders') else None,
+                'rec_crystal_correction': int(rec_match.group('rf_correction_manual')) if rec_match.group('rf_correction_manual') else 0,
+                'rec_crystal_correction_fine': float(rec_match.group('rf_correction_automatic')) if rec_match.group('rf_correction_automatic') else 0.0,
+                'rec_input_noise': float(rec_match.group('signal')) if rec_match.group('signal') else None,
+                'senders_signal': float(rec_match.group('senders_signal')) if rec_match.group('senders_signal') else None,
+                'senders_messages': float(rec_match.group('senders_messages')) if rec_match.group('senders_messages') else None,
+                'good_senders_signal': float(rec_match.group('good_senders_signal')) if rec_match.group('good_senders_signal') else None,
+                'good_senders': float(rec_match.group('good_senders')) if rec_match.group('good_senders') else None,
+                'good_and_bad_senders': float(rec_match.group('good_and_bad_senders')) if rec_match.group('good_and_bad_senders') else None}
     else:
         return None